task_Re/requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
pip==23.2.1
pytz==2024.1
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2023.4